Move `any` from `cetl::pf17` → just `cetl` namespace.
`cetl::pf17` will have very close to `std::any` specialization.
